[#101731950] this should address the race condition
we have seen a race condition intermittently fail
our tests, when using the -race flag.

our code changes where we dive into a
sub goroutine so that we are not passing
uninstantiated objects... at least thats
what should happen.

also modified the callbacks in the test
to use a single fakeCallback, for
easier use