[finished #101296514] we are now able to make a lease POST
we are now able to make a POST lease call
w/ a lease for a 2c.small and have it
take action on the procurement.

we also will have a task to follow its
status, as well as a managerial task
to track the availablility of the leased
items internally.