[#101296512] polling mechanism is now hooked up.
we have a little more clarity and functionality around the
polling mechanism.. we are not locking, we are pushing back the expiration
time, so its not checked again, this should give us time to
run any analysis on it. if the run fails or the service
fails, it will get picked up again when the timeout
expires, in the form of the expires field value (epoch nano).

if its 0 then it has already expired and should not be looked at.

begging for integrations of some sort, and needs to be tied into
the actual endpoint.... more to come